requirements.txt
fastapi
python-a2s
